move auto wrappers to common/auto directory
