+++ src/client_handler.cpp
@@ -1,7 +1,7 @@
 #include "proto/server_message.pb.h"
 
 #include "client_handler.hpp"
-#include "common/auto_lock.hpp"
+#include "common/autos/auto_lock.hpp"
 
 ClientHandler::ClientHandler(std::shared_ptr<Client> client, std::shared_ptr<RulesManager> rules_manager)
     : _client(std::move(client)), _rules_manager(std::move(rules_manager))
