+++ src/common/autos/auto_lock.cpp
