+++ src/common/server_socket.cpp
@@ -2,9 +2,9 @@
 #include <sys/socket.h>
 #include <unistd.h>
 
-#include "../exceptions.hpp"
 #include "server_socket.hpp"
 #include "os_utils.hpp"
+#include "unix.hpp"
 
 ServerSocket::ServerSocket(const uint32_t port) : _socket_fd(
                                                       os::covered_call(::socket, AF_INET,
